Trait linxal::eigenvalues::general::Eigen

pub trait Eigen: Sized + Clone {
    type Eigv;
    type Solution;
    fn compute_into<D>(mat: ArrayBase<D, Ix2>, compute_left: bool, compute_right: bool) -> Result<Self::Solution, EigenError> where D: DataOwned<Elem=Self> + DataMut<Elem=Self>;

    fn compute<D>(mat: &ArrayBase<D, Ix2>, compute_left: bool, compute_right: bool) -> Result<Self::Solution, EigenError> where D: Data<Elem=Self> { ... }
}

Associated Types

type Eigv

type Solution

Required Methods

fn compute_into<D>(mat: ArrayBase<D, Ix2>, compute_left: bool, compute_right: bool) -> Result<Self::Solution, EigenError> where D: DataOwned<Elem=Self> + DataMut<Elem=Self>

Return the eigenvalues and, optionally, the left and/or right eigenvectors of a general matrix.

The entries in the input matrix mat are modified when calculating the eigenvalues.

Provided Methods

fn compute<D>(mat: &ArrayBase<D, Ix2>, compute_left: bool, compute_right: bool) -> Result<Self::Solution, EigenError> where D: Data<Elem=Self>

Return the eigenvvalues and, optionally, the eigenvectors of a general matrix.