Struct linxal::eigenvalues::types::Solution

pub struct Solution<IV, EV> {
    pub values: Array<EV, Ix1>,
    pub left_vectors: Option<Array<IV, Ix2>>,
    pub right_vectors: Option<Array<IV, Ix2>>,
}

Solution to an eigenvalue problem.

Contains thte eigenvalues and, optionally, the left and/or right eigenvectors of the solution. For symmetric problems, the eigenvectors are placed in right_eigenvectors.

Fields
values: Array<EV, Ix1>

Eigenvalues of the matrix.

left_vectors: Option<Array<IV, Ix2>>

Eigenvectors for the left-eigenvalue problem ($xH A = \lambda xH$)

right_vectors: Option<Array<IV, Ix2>>

Eigenvectors for the right-eigenvalue problem ($Ax = \lambda x$)

For symmetric eigenvalue problems, the eigenvectors will be stored in right_vectors.