Struct linxal::svd::types::SVDSolution

pub struct SVDSolution<IV: Sized, SV: Sized> {
    pub values: Array<SV, Ix1>,
    pub left_vectors: Option<Array<IV, Ix2>>,
    pub right_vectors: Option<Array<IV, Ix2>>,
}

A solution to the singular value decomposition.

A singular value decomposition solution includes singular values and, optionally, the left and right singular vectors, stored as a mtrix.

Fields
values: Array<SV, Ix1>

Singular values of the matrix.

Singular values, which are guaranteed to be non-negative reals, are returned in descending order.

left_vectors: Option<Array<IV, Ix2>>

The matrix U of left singular vectors.

right_vectors: Option<Array<IV, Ix2>>

The matrix Vt of singular vectors.

The transpose of V is stored, not V itself.